// bga-cards.d.ts
type SideOrAngle = 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | 'top' | 'bottom' | 'left' | 'right'
// ...
interface SelectionStyle {
	/**
	 * The class to apply to this selection style. Use class from manager is unset in the card stock.
	 */
	class?: string | null
	/**
	 * The outline size. Default 5px;
	 */
	outlineSize?: number
	/**
	 * The outline color. Default 'orange' for selectable dashes, 'blueviolet' for selected line.
	 */
	outlineColor?: string
}
interface CardStockSettings<T> {
	/**
	 * Indicate the card sorting (unset means no sorting, new cards will be added at the end).
	 * For example, use `sort: sortFunction('type', '-type_arg')` to sort by type then type_arg (in reversed order if prefixed with `-`).
	 * Be sure you typed the values correctly! Else '11' will be before '2'.
	 */
	sort?: SortFunction<T>
	/**
	 * The filter on card click event. Use setting from manager is unset.
	 */
	cardClickEventFilter?: CardClickEventFilter
	/**
	 * The style to apply to selectable cards. Use style from manager is unset.
	 */
	selectableCardStyle?: SelectionStyle
	/**
	 * The style to apply to selectable cards. Use style from manager is unset.
	 */
	unselectableCardStyle?: SelectionStyle
	/**
	 * The style to apply to selected cards. Use style from manager is unset.
	 */
	selectedCardStyle?: SelectionStyle
	/**
	 * The style to apply to the last played card. Default to class 'bga-cards_last-played-card'.
	 */
	lastPlayedCardStyle?: SelectionStyle
	/**
	 * Show a card counter on the stock. Not visible if unset.
	 */
	counter?: CardCounterSettings
	/**
	 * Say if a given card should be placed on this stock, based on card properties.
	 * For example, every card with location === 'hand' and location_arg == this.player_id should go to the current player hand.
	 * If unset, all cards on this stock must be added manually on this stock.
	 *
	 * @param card the card to place on a stock
	 * @returns true if the card should be placed on this stock.
	 */
	autoPlace?: (card: T) => boolean
}
interface AddCardSettings extends SlideAnimationSettings {
	/**
	 * The stock to take the card. It will automatically remove the card from the other stock.
	 */
	fromStock?: CardStock<any>
	/**
	 * The element to move the card from.
	 */
	fromElement?: HTMLElement
	/**
	 * Card side at the beginning of the animation. Default 'auto' to isCardVisible. Ignored if the card already exists.
	 */
	initialSide?: 'auto' | 'front' | 'back'
	/**
	 * Card side at the end of the animation. Default to initialSide.
	 */
	finalSide?: 'auto' | 'front' | 'back'
	forceToElement?: HTMLElement
	/**
	 * Force card position. Default to end of list. Do not use if sort is defined, as it will override it.
	 */
	index?: number
	/**
	 * Set if the card is selectable. Default is true, but will be ignored if the stock is not selectable.
	 */
	selectable?: boolean
	/**
	 * Indicates if we add a fade in effect when adding card (if it comes from an invisible or abstract element).
	 */
	fadeIn?: boolean
	/**
	 * For counters.
	 * Indicate if the card count is automatically updated when a card is added or removed. Default true.
	 */
	autoUpdateCardNumber?: boolean
}
interface RemoveCardSettings {
	slideTo?: HTMLElement
	fadeOut?: boolean
	/**
	 * For counters.
	 * Indicate if the card count is automatically updated when a card is added or removed. Default true.
	 */
	autoUpdateCardNumber?: boolean
}
type CardSelectionMode = 'none' | 'single' | 'multiple'
// ...
interface DeckSettings<T> extends CardStockSettings<T> {
	/**
	 * Indicate the current top card.
	 */
	topCard?: T
	/**
	 * Indicate the current number of cards in the deck (default 0).
	 */
	cardNumber?: number
	/**
	 * Indicate if the card count is automatically updated when a card is added or removed.
	 */
	autoUpdateCardNumber?: boolean
	/**
	 * Indicate if the cards under the new top card must be removed (to forbid players to check the content of the deck with Inspect). Default true.
	 */
	autoRemovePreviousCards?: boolean
	/**
	 * Indicate the thresholds to add 1px to the thickness of the pile. Default [0, 2, 5, 10, 20, 30].
	 */
	thicknesses?: number[]
	/**
	 * Shadow direction. Default 'bottom-right'.
	 */
	shadowDirection?: SideOrAngle
	/**
	 * A generator of fake cards, to generate decks top card automatically.
	 * Default is manager `fakeCardGenerator` method.
	 *
	 * @param deckId the deck id
	 * @return the fake card to be generated (usually, only informations to show back side)
	 */
	fakeCardGenerator?: (deckId: string) => T
}
interface AddCardToDeckSettings extends AddCardSettings {
	/**
	 * Indicate if the cards under the new top card must be removed (to forbid players to check the content of the deck with Inspect). Default true.
	 */
	autoRemovePreviousCards?: boolean
}
interface RemoveCardFromDeckSettings extends RemoveCardSettings {
	/**
	 * Indicate if the card count is automatically updated when a card is added or removed.
	 */
	autoUpdateCardNumber?: boolean
}
interface ShuffleAnimationSettings<T> {
	/**
	 * Number of cards used for the animation (will use cardNumber is inferior to this number).
	 * Default: 10.
	 */
	animatedCardsMax?: number
	/**
	 * Card generator for the animated card. Should only show the back of the cards.
	 * Default if fakeCardGenerator from Deck (or Manager if unset in Deck).
	 */
	fakeCardSetter?: (card: T, index: number) => void
	/**
	 * Time to wait after shuffle, in case it is chained with other animations, to let the time to understand it's 2 different animations.
	 * Default is 500ms.
	 */
	pauseDelayAfterAnimation?: number
}
/**
 * Stock to represent a deck. (pile of cards, with a fake 3d effect of thickness).
 */
declare class Deck<T> extends CardStock<T> {
	protected manager: CardManager<T>
	protected element: HTMLElement
	protected cardNumber: number
	protected autoUpdateCardNumber: boolean
	protected autoRemovePreviousCards: boolean
	protected fakeCardGenerator?: (deckId: string) => T
	protected thicknesses: number[]
	constructor(manager: CardManager<T>, element: HTMLElement, settings: DeckSettings<T>)
	/**
	 * Get the the cards number.
	 *
	 * @returns the cards number
	 */
	getCardNumber(): number
	/**
	 * Set the the cards number.
	 *
	 * @param cardNumber the cards number
	 * @param topCard the deck top card. If unset, will generated a fake card (default). Set it to null to not generate a new topCard.
	 */
	setCardNumber(cardNumber: number, topCard?: T | null | undefined): Promise<boolean>
	addCard(card: T, settings?: AddCardToDeckSettings): Promise<boolean>
	cardRemoved(card: T, settings?: RemoveCardFromDeckSettings): void
	removeAll(settings?: RemoveCardFromDeckSettings): Promise<boolean>
	getTopCard(): T | null
	/**
	 * Shows a shuffle animation on the deck
	 *
	 * @param settings a `ShuffleAnimationSettings` object
	 * @returns promise when animation ends
	 */
	shuffle(settings?: ShuffleAnimationSettings<T>): Promise<boolean>
	protected getFakeCard(): T
	/**
	 * Returns the card count in the deck (what the player think there is, for decks, the real number of cards for all visible card stocks).
	 *
	 * @returns the number of card in the stock
	 */
	getCardCount(): number
}
// ...
declare class LineStock<T> extends CardStock<T> {
	protected manager: CardManager<T>
	protected element: HTMLElement
	/**
	 * @param manager the card manager
	 * @param element the stock element (should be an empty HTML Element)
	 * @param settings a `LineStockSettings` object
	 */
	constructor(manager: CardManager<T>, element: HTMLElement, settings?: LineStockSettings<T>)
}

interface SlotStockSettings<T> extends LineStockSettings<T> {
	/**
	 * The ids for the slots (can be number or string)
	 */
	slotsIds: SlotId[]
	/**
	 * The classes to apply to each slot
	 */
	slotClasses?: string[]
	/**
	 * How to place the card on a slot automatically
	 */
	mapCardToSlot: (card: T) => SlotId
	/**
	 * The style to apply to selectable slots. Use class from manager is unset.
	 */
	selectableSlotStyle?: SelectionStyle
	/**
	 * The style to apply to unselectable slots. Use class from manager is unset.
	 */
	unselectableSlotStyle?: SelectionStyle
	/**
	 * The style to apply to selected slots. Use class from manager is unset.
	 */
	selectedSlotStyle?: SelectionStyle
}
type SlotId = number | string
interface AddCardToSlotSettings extends AddCardSettings {
	/**
	 * The slot to place the card on.
	 */
	slot?: SlotId
}
/**
 * A stock with fixed slots (some can be empty)
 */
declare class SlotStock<T> extends LineStock<T> {
	protected manager: CardManager<T>
	protected element: HTMLElement
	protected slotsIds: SlotId[]
	protected slots: HTMLDivElement[]
	protected slotClasses: string[]
	protected mapCardToSlot: (card: T) => SlotId
	protected selectedSlots: SlotId[]
	protected slotSelectionMode: CardSelectionMode
	/**
	 * Called when the slot selection change. Returns the selection.
	 *
	 * selection: the selected SlotId of the stock
	 * lastChange: the last change on selection slot (can be selected or unselected)
	 */
	onSlotSelectionChange?: (selection: SlotId[], lastChange: SlotId | null) => void
	/**
	 * Called when slot selection change. Returns the clicked slot.
	 *
	 * slot: the clicked slot (can be selected or unselected)
	 */
	onSlotClick?: (slotId: SlotId) => void
	/**
	 * @param manager the card manager
	 * @param element the stock element (should be an empty HTML Element)
	 * @param settings a `SlotStockSettings` object
	 */
	constructor(manager: CardManager<T>, element: HTMLElement, settings: SlotStockSettings<T>)
	protected setSlotSelectionStyleOverrides(element: HTMLElement): void
	protected createSlot(slotId: SlotId): void
	/**
	 * Add a card to the stock.
	 *
	 * @param card the card to add
	 * @param settings a `AddCardToSlotSettings` object
	 * @returns the promise when the animation is done (true if it was animated, false if it wasn't)
	 */
	addCard(card: T, settings?: AddCardToSlotSettings): Promise<boolean>
	getSlotsIds(): SlotId[]
	/**
	 * Change the slots ids. Will empty the stock before re-creating the slots.
	 *
	 * @param slotsIds the new slotsIds. Will replace the old ones.
	 */
	setSlotsIds(slotsIds: SlotId[]): void
	removeSlot(slotId: SlotId): void
	/**
	 * Add new slots ids. Will not change nor empty the existing ones.
	 *
	 * @param newSlotsIds the new slotsIds. Will be merged with the old ones.
	 */
	addSlotsIds(newSlotsIds: SlotId[]): void
	/**
	 * @returns the style to apply to selectable slots. Use style from manager is unset.
	 */
	getSelectableSlotStyle(): SelectionStyle
	/**
	 * @returns the style to apply to selectable slots. Use style from manager is unset.
	 */
	getUnselectableSlotStyle(): SelectionStyle
	/**
	 * @returns the style to apply to selected slots. Use style from manager is unset.
	 */
	getSelectedSlotStyle(): SelectionStyle
	protected canAddCard(card: T, settings?: AddCardToSlotSettings): boolean
	/**
	 * Swap cards inside the slot stock.
	 *
	 * @param cards the cards to swap
	 * @param settings for `updateInformations` and `selectable`
	 */
	swapCards(cards: T[], settings?: AddCardSettings): any
	/**
	 * Set if the stock slot are selectable, and if yes if it can be multiple.
	 * If set to 'none', it will unselect all selected slots.
	 *
	 * @param selectionMode the selection mode
	 * @param selectableSlots the selectable slats (all if unset). Calls `setSelectableSlots` method
	 */
	setSlotSelectionMode(selectionMode: CardSelectionMode, selectableSlots?: SlotId[]): void
	removeSlotSelectionClasses(slotId: SlotId): void
	removeSlotSelectionClassesFromElement(slotElement: HTMLElement): void
	protected setSelectableSlot(slotId: SlotId, selectable: boolean): void
	/**
	 * Set the selectable class for each slot.
	 *
	 * @param slotIds the selectable slots. If unset, all slots are marked selectable. Default unset.
	 */
	setSelectableSlots(slotIds?: SlotId[]): void
	/**
	 * Set selected state to a slot.
	 *
	 * @param slotId the slot to select
	 */
	selectSlot(slotId: SlotId): void
	/**
	 * Set unselected state to a slot.
	 *
	 * @param slotId the slot to unselect
	 */
	unselectSlot(slotId: SlotId): void
	/**
	 * Select all slots
	 */
	selectAllSlots(): void
	/**
	 * Unselect all slots
	 */
	unselectAllSlots(): void
	/**
	 * @returns the selected slots
	 */
	getSlotSelection(): SlotId[]
	/**
	 * @returns if the slot is selectd
	 */
	isSlotSelected(slotId: SlotId): boolean
}